chore: refactor `QrCodeDisplay` & `startQrPresentment`
